// utils/enums.ts
export enum StatusCode {
  OK = 200,
  CREATED = 201,
  BAD_REQUEST = 400,
}

export enum Teacher_Roles {
  HEAD_TEACHER = "head teacher",
  SUBJECT_TEACHER = "subject of teacher",
  ASSISTANT_TEACHER = "assistant of a teacher",
}

export enum staffDuty {
  VICE_PRINCIPAL = "vice principal, ",
  PRINCIPAL = "principal",
  TEACHER = "teacher",
}